Fix dentist create method
