+++ angular/dental/src/app/patient/alert/alert.service.ts
@@ -1,7 +1,6 @@
 import { Injectable } from '@angular/core';
-import { BaseAuthorizedService } from '../../shared/auth/base-authorized.service';
 
 @Injectable()
-export class AlertService extends BaseAuthorizedService{
+export class AlertService{
 
 }
